Move font/text related stuff to an extension. Split utils into multiple files. Renamed buffer_pool and atlas_pool to buffer and atlas. New text renderer, more optimal than using the render2d pipeline. Replaced buffer::ref by new suballoc type, that may be used as a suballoc of a buffer or another alloc (like a suballoc in a renderer buffer). Removed most setter from types, encourage using updators directly (some setters available in suballoc, but based on updator). Commit to vulkan, and expose way more of its types in the API, no support for other API will ever happen, this is a desktop and linux only library. Remove usages of cstdint usage for consistency, and added a clang-format config.

+++ ext/render2d/spv/render2d.frag
@@ -1,10 +1,10 @@
-#version 450
+#version 460
 #extension GL_ARB_separate_shader_objects : enable
 #extension GL_EXT_nonuniform_qualifier : require
 
 layout(binding = 1) uniform sampler2D uni_samplers[4];
 
-layout(location = 0) in flat uvec4 in_box; // pos_box
+layout(location = 0) in flat vec4 in_box; // pos_box
 layout(location = 1) in flat uvec4 in_params; // (radius, smoothness, atlas_page, gradient)
 layout(location = 2) in vec4 in_col;
 layout(location = 3) in vec2 in_uv;
